Wrote script for compiling latex files.
Main use of script is for emacs to use when exporting org-mode files. Instead of
running the set TeX product an arbitrary number of times (sometimes runs too
many times, which is inefficient, and sometimes too few, which is
unhelpful/dangerous), use mklatex to fully compile in an efficient manner.

The script can also be called from the command line, so have included a zsh
completion script too.